requirements.txt
# Core ML and retrieval
torch>=2.0.0
transformers>=4.40.0
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2

# Vision-Language Model support
accelerate>=0.29.0
bitsandbytes>=0.43.0
Pillow>=10.0.0

# Document processing and scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
html2text>=2020.1.16
tqdm>=4.66.0

# Evaluation
ragas>=0.1.7
datasets>=2.18.0
nltk>=3.8.1
rouge-score>=0.1.2
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.0.0

# Answer generation via API
openai>=1.14.0
anthropic>=0.20.0

# Demo interface
gradio>=4.25.0

# Config and utilities
PyYAML>=6.0.1
python-dotenv>=1.0.0
jsonlines>=4.0.0
rich>=13.7.0
click>=8.1.7

# Experiment tracking
matplotlib>=3.8.0
seaborn>=0.13.0
